[lib] implement a translation toolchain
Introduce a new `@/lib/translators/toolchain` module that exposes
a `getTranslator` function. This will help to register several backend
translators and pick up one of them.

Move out the mongo implementation in its own `@/lib/translators/mongo`
module and register it in the new toolchain framework. `pipe2mongo`
shouldn't be called directly, you should use the following code instead:

```typescript
import { getTranslator } from '@/lib/translators/toolchain';

const mongo36translator = getTranslator('mongo36');
const pipeline: Array<PipelineStep> = [{ name: 'domain', domain: 'test_cube' }];
const querySteps = mongo36translator(pipeline);
```
